+++ modules/apps/portal-vulcan/portal-vulcan-api/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Portal Vulcan API
+Bundle-SymbolicName: com.liferay.portal.vulcan.api
+Bundle-Version: 1.0.0
+Export-Package:\
+	com.liferay.portal.vulcan.collection,\
+	com.liferay.portal.vulcan.context
