+++ modules/apps/portal-vulcan/portal-vulcan-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Portal Vulcan Implementation
+Bundle-SymbolicName: com.liferay.portal.vulcan.impl
+Bundle-Version: 1.0.0
